+++ app/Propstack/Imports/v1/Objects.php
@@ -10,7 +10,6 @@ namespace ConnectorForPropstack\Propstack\Imports\v1;
 // prevent direct access.
 defined( 'ABSPATH' ) || exit;
 
-use Error;
 use ConnectorForPropstack\Plugin\Helper;
 use ConnectorForPropstack\Plugin\Languages;
 use ConnectorForPropstack\Plugin\Log;
@@ -382,34 +381,36 @@ class Objects extends Import_Base {
 				// report the success.
 				$process_handler->set_message( $this->get_success_dialog_config() );
 			}
-		} catch ( Error $e ) {
+		} catch ( \Throwable $e ) {
 			// log this event.
 			Log::get_instance()->add( __( 'Following error occurred during the import of objects via API v2:', 'connector-for-propstack' ) . '<br>' . __( 'Message:', 'connector-for-propstack' ) . '<code>' . $e->getMessage() . '</code><br>' . __( 'Code:', 'connector-for-propstack' ) . '<code>' . $e->getCode() . '</code><br>' . __( 'File:', 'connector-for-propstack' ) . '<code>' . $e->getFile() . '</code><br>' . __( 'Line:', 'connector-for-propstack' ) . '<code>' . $e->getLine() . '</code>', 'error', 'import' );
 
 			// show hint.
 			/* translators: %1$s will be replaced by a URL. */
 			$this->add_error( 'propstack_object_import_error', sprintf( __( 'Error occurred. Check <a href="%1$s">the log</a> for details.', 'connector-for-propstack' ), esc_url( Settings::get_instance()->get_url( 'propstack_connector_logs' ) ) ) );
+		} finally {
+
+			/**
+			 * Run additional tasks after any import of objects.
+			 *
+			 * @since 1.0.0 Available since 1.0.0.
+			 *
+			 * @param Objects $instance The import object.
+			 */
+			do_action( 'cfprop_import_object_after', $instance );
+
+			// log the errors.
+			$this->save_errors_in_log();
+
+			// add a log entry if debug is enabled.
+			if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
+				Log::get_instance()->add( __( 'Import of objects has been ended.', 'connector-for-propstack' ), 'info', 'import' );
+			}
+
+			// update the running marker.
+			$process_handler->set_running( 0 );
+			update_option( CFPROP_IMPORT_RUNNING, 0 );
 		}
-
-		/**
-		 * Run additional tasks after any import of objects.
-		 *
-		 * @since 1.0.0 Available since 1.0.0.
-		 * @param Objects $instance The import object.
-		 */
-		do_action( 'cfprop_import_object_after', $instance );
-
-		// log the errors.
-		$this->save_errors_in_log();
-
-		// add a log entry if debug is enabled.
-		if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-			Log::get_instance()->add( __( 'Import of objects has been ended.', 'connector-for-propstack' ), 'info', 'import' );
-		}
-
-		// update the running marker.
-		$process_handler->set_running( 0 );
-		update_option( CFPROP_IMPORT_RUNNING, 0 );
 	}
 
 	/**