+++ changelog.md
@@ -4,6 +4,19 @@
 
 ### Changed
 
+- Optimized capability checks
+- Optimized adding lines in CSV-export for logs
+- Scheduled import for objects will now use API v2 if set
+
+### Fixed
+
+- Fixed a logical error during the check of the object type on any object to import
+- Fixed a wrong filter for property types during import of objects with API v1
+
+## [1.0.1] - 13.07.2026
+
+### Changed
+
 - Optimized loading of object types
 - Use global helper to nat sort arrays
 - Updated crypt library to 2.0.1